use MockServer for RetryTest
